🔧 fix: VoteProperties 및 관련 클래스에서 가중치 타입을 Long에서 int로 변경
- VoteProperties 클래스의 가중치 필드를 Long에서 int로 수정
- SimilarityVoteSummaryEntity의 updateFinalDecision 메서드에 최소 총 가중치 매개변수 추가
- SimilarityVoteSummaryRepository의 isVoteCompleted 메서드에서 Long 타입을 int로 변경
- VotingService에서 updateFinalDecision 호출 시 임계치 기반으로 수정

가중치 처리의 일관성을 높이고 성능을 개선하기 위한 변경 사항입니다.

+++ src/main/java/hanium/modic/backend/common/property/property/VoteProperties.java
@@ -10,9 +10,9 @@ import lombok.Setter;
 @ConfigurationProperties(prefix = "vote.similarity")
 public class VoteProperties {
 
-	private Long minTotalWeight;
-	private Long aiVoteWeight;
-	private Long humanVoteWeight;
+	private int minTotalWeight;
+	private int aiVoteWeight;
+	private int humanVoteWeight;
 	private Boolean enableAiAssessment;
-	private Long maxVotesPerUserPerDay;
+	private int maxVotesPerUserPerDay;
 }
